Introduced a base DamageCalculator
