+++ openquake/commonlib/calculators/scenario_damage.py
@@ -64,61 +64,13 @@ def scenario_damage(riskinputs, riskmodel, rlzs_assoc, monitor):
 
 
 @base.calculators.add('scenario_damage')
-class ScenarioDamageCalculator(base.RiskCalculator):
+class ScenarioDamageCalculator(base.DamageCalculator):
     """
     Scenario damage calculator
     """
     hazard_calculator = 'scenario'
     core_func = scenario_damage
     result_kind = 'damages_by_key'
-
-    def pre_execute(self):
-        """
-        Get the GMFs and build the riskinputs.
-        """
-        super(ScenarioDamageCalculator, self).pre_execute()
-        if 'gmfs' in self.oqparam.inputs:  # from file
-            gmfs = self.read_gmfs_from_csv()
-        else:  # from rupture
-            gmfs = self.compute_gmfs()
-        self.riskinputs = self.build_riskinputs(gmfs)
-
-    def compute_gmfs(self):
-        """
-        :returns: riskinputs
-        """
-        logging.info('Computing the GMFs')
-        haz_out, hcalc = base.get_hazard(self)
-        gmfs_by_trt_gsim = calc.expand(
-            haz_out['gmfs_by_trt_gsim'], haz_out['sites'])
-
-        logging.info('Preparing the risk input')
-        self.rlzs_assoc = haz_out['rlzs_assoc']
-        return gmfs_by_trt_gsim
-
-    def read_gmfs_from_csv(self):
-        """
-        :returns: riskinputs
-        """
-        logging.info('Reading hazard curves from CSV')
-        sitecol, gmfs_by_imt = readinput.get_sitecol_gmfs(self.oqparam)
-
-        # filter the hazard sites by taking the closest to the assets
-        with self.monitor('assoc_assets_sites'):
-            self.sitecol, self.assets_by_site = self.assoc_assets_sites(
-                sitecol)
-
-        # reduce the gmfs matrices to the filtered sites
-        for imt in gmfs_by_imt:
-            gmfs_by_imt[imt] = gmfs_by_imt[imt][self.sitecol.indices]
-
-        num_assets = sum(len(assets) for assets in self.assets_by_site)
-        num_sites = len(self.sitecol)
-        logging.info('Associated %d assets to %d sites', num_assets, num_sites)
-
-        logging.info('Preparing the risk input')
-        self.rlzs_assoc = riskinput.FakeRlzsAssoc(num_rlzs=1)
-        return {(0, 'FromCsv'): gmfs_by_imt}
 
     def post_execute(self, result):
         """
